Feat: implement of more errors validations in controllers

+++ src/controllers/ticketController.js
@@ -6,6 +6,11 @@ class TicketController{
     async findAllTickets(req,res){
         try{
             const findAll = await TicketService.findAllTickets();
+
+                if(findAll.length === 0){
+                    res.status(404).json({
+                        message: "Empty tickets list"
+                })}
             res.status(200).json({
                 Tickets: findAll
             })
@@ -20,12 +25,18 @@ class TicketController{
         try{
             const { idTicket } = req.params;
             const ticketData = await TicketService.findTicketByPk(idTicket);
-            res.status(200).json({
-                Ticket: ticketData
-            })
-        }catch(e){
-            res.status(400).json({
-                Error: e.message
+
+                if(!ticketData){
+                    res.status(404).json({
+                        message: "Ticket not found"
+                    })
+                }else{
+                    res.status(200).json({
+                        Ticket: ticketData
+                })}
+                }catch(e){
+                    res.status(400).json({
+                        Error: e.message
             })
         }
     }
@@ -47,13 +58,19 @@ class TicketController{
     async deleteTicket(req,res){
         try{
             const { idTicket } = req.params;
-            const deletedTicket = await TicketService.deleteTicket(idTicket);
-            res.status(200).json({
-                Deleted: deletedTicket
-            })
-        }catch(e){
-            res.status(400).json({
-                Error: e.message
+            
+            if(!idTicket){
+                res.status(404).json({
+                    message: "Ticket not found"
+                })}
+            else{
+                const deletedTicket = await TicketService.deleteTicket(idTicket);
+                res.status(200).json({
+                    Deleted: deletedTicket
+            })}
+            }catch(e){
+                res.status(400).json({
+                    Error: e.message
             })
         }
     }
@@ -62,14 +79,20 @@ class TicketController{
         try{
             const { idTicket } = req.params;
             const ticketData = req.body;
+
+            if(!idTicket){
+                res.status(404).json({
+                    message: "Ticket not found"
+                })}
+            else{
             await TicketService.updateTicket(idTicket, ticketData);
             const result = await TicketService.findTicketByPk(idTicket);
-            res.status(200).json({
-                Updated: result
-             });
-        }catch(e){
-            res.status(400).json({
-                Error: e.message
+                res.status(200).json({
+                    Updated: result
+             })}
+            }catch(e){
+                res.status(400).json({
+                    Error: e.message
             })
         }
     }
